+++ module-array/src/main/java/ch/hsr/adv/ui/array/logic/domain/ArrayRelation.java
@@ -1,8 +1,8 @@
 package ch.hsr.adv.ui.array.logic.domain;
 
+import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
+import ch.hsr.adv.commons.core.logic.domain.styles.presets.ADVDefaultLineStyle;
 import ch.hsr.adv.ui.core.logic.domain.ADVRelation;
-import ch.hsr.adv.ui.core.logic.domain.styles.ADVStyle;
-import ch.hsr.adv.ui.core.logic.domain.styles.presets.ADVDefaultLineStyle;
 
 /**
  * Relation between two {@link ArrayElement}s
