+++ module-graph/src/main/java/ch/hsr/adv/ui/graph/logic/domain/GraphRelation.java
@@ -1,8 +1,8 @@
 package ch.hsr.adv.ui.graph.logic.domain;
 
 
+import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
 import ch.hsr.adv.ui.core.logic.domain.ADVRelation;
-import ch.hsr.adv.ui.core.logic.domain.styles.ADVStyle;
 
 /**
  * Object adapter. Maps an ADVEdge to an ADVRelation.
