+++ ui-core/src/main/java/ch/hsr/adv/ui/core/logic/domain/ADVElement.java
@@ -1,6 +1,6 @@
 package ch.hsr.adv.ui.core.logic.domain;
 
-import ch.hsr.adv.ui.core.logic.domain.styles.ADVStyle;
+import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
 
 /**
  * Represents a generic elements which is displayed on the ADV UI.
