--- ui-core/src/main/java/ch/hsr/adv/ui/core/logic/util/ADVException.java
@@ -1,15 +1,0 @@
-package ch.hsr.adv.ui.core.logic.util;
-
-/**
- * The main exception class to be used for ADVCore projects.
- */
-public class ADVException extends Exception {
-    /**
-     * Constructs an ADVException with the given message.
-     *
-     * @param message the error messsage
-     */
-    public ADVException(String message) {
-        super(message);
-    }
-}
