+++ ui-core/src/main/java/ch/hsr/adv/ui/core/logic/util/ADVParseException.java
@@ -3,7 +3,7 @@ package ch.hsr.adv.ui.core.logic.util;
 /**
  * To be thrown for any exception to do with parsing of a session, snapshot etc.
  */
-public class ADVParseException extends ADVException {
+public class ADVParseException extends Exception {
 
     public ADVParseException(String message) {
         super(message);
