+++ ui-core/src/main/java/ch/hsr/adv/ui/core/presentation/widgets/CurvedLabeledEdge.java
@@ -1,6 +1,6 @@
 package ch.hsr.adv.ui.core.presentation.widgets;
 
-import ch.hsr.adv.ui.core.logic.domain.styles.ADVStyle;
+import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
 import javafx.geometry.Point2D;
 import javafx.scene.shape.CubicCurve;
 import org.slf4j.Logger;
