+++ ui-core/src/main/java/ch/hsr/adv/ui/core/presentation/widgets/LabeledEdge.java
@@ -1,6 +1,6 @@
 package ch.hsr.adv.ui.core.presentation.widgets;
 
-import ch.hsr.adv.ui.core.logic.domain.styles.ADVStyle;
+import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
 import ch.hsr.adv.ui.core.presentation.util.StyleConverter;
 import javafx.beans.binding.Bindings;
 import javafx.beans.binding.DoubleBinding;
