+++ ui-core/src/test/java/ch/hsr/adv/ui/core/logic/domain/styles/ADVValueStyleTest.java
@@ -1,5 +1,7 @@
 package ch.hsr.adv.ui.core.logic.domain.styles;
 
+import ch.hsr.adv.commons.core.logic.domain.styles.ADVStrokeStyle;
+import ch.hsr.adv.commons.core.logic.domain.styles.ADVValueStyle;
 import com.google.inject.Inject;
 import org.jukito.JukitoRunner;
 import org.junit.Assert;
